# docs/swagger.yaml
basePath: /api
host: 114.132.40.112:9502
info:
  contact:
    email: pltrueover@gmail.com
    name: go-core
    url: https://im.pltrue.top
  description: go-im 服务端
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: go-im  接口文档
  version: "2.0"
paths:
  /AddSession:
    post:
      description: 添加会话信息
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 好友id或者群聊id
        in: formData
        name: f_id
        required: true
        type: string
      - description: 会话类型 1.单聊 2.群聊
        in: formData
        name: channel_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 添加会话信息
      tags:
      - 添加会话信息
...
  /CreateGroup:
    post:
      description: 创建群聊
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 群聊名称
        in: formData
        name: group_name
        required: true
        type: string
      - description: 群聊用户
        in: formData
        name: user_id
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 创建群聊
      tags:
      - 创建群聊
  /DelSession:
    post:
      description: 删除会话信息
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 好友id或者群聊id
        in: formData
        name: f_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 删除会话信息
      tags:
      - 删除会话信息
...
  /GetGroupDetails:
    get:
      description: 获取群聊详情
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 群聊id
        in: query
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 获取群聊详情
      tags:
      - 获取群聊详情
  /GetGroupList:
    get:
      description: 获取群聊列表
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 获取群聊列表
      tags:
      - 获取群聊列表
...
  /GetSessionList:
    get:
      description: 获取会话列表
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 获取会话列表
      tags:
      - 获取会话列表
...
  /JoinGroup:
    post:
      description: 添加用户到指定群聊
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 群聊id
        in: formData
        name: group_id
        required: true
        type: string
      - description: 用户id
        in: formData
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 添加用户到指定群聊
      tags:
      - 添加用户到指定群聊
...
  /RemoveGroup:
    post:
      description: 删除群聊
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 群聊id
        in: formData
        name: group_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 删除群聊
      tags:
      - 删除群聊
  /RemovedUserFromGroup:
    post:
      description: 移除群聊用户
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 群聊id
        in: formData
        name: group_id
        required: true
        type: string
      - description: 用户id
        in: formData
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 移除群聊用户
      tags:
      - 移除群聊用户
...
  /SetSessionTop:
    post:
      description: 会话置顶
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: 会话id
        in: formData
        name: session_id
        required: true
        type: string
      - description: 0 正常 1置顶
        in: formData
        name: top_status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: 会话置顶
      tags:
      - 会话置顶
...
swagger: "2.0"